+++ src/components/i18n/i18n-provider.tsx
@@ -1,6 +1,4 @@
-'use client';
-
-import i18n from '../../../utils/i18n';
+import i18n from '../../utils/i18n';
 import { I18nextProvider } from 'react-i18next';
 
 export default function I18nProvider({ children }: { children: React.ReactNode }) {
