+++ src/pages/_app.tsx
@@ -1,0 +1,41 @@
+import type { AppProps } from 'next/app';
+import Head from 'next/head';
+import { Geist, Geist_Mono, Roboto } from "next/font/google";
+import I18nProvider from "../components/i18n/i18n-provider";
+
+import "../globals.css";
+
+const geistSans = Geist({
+  variable: "--font-geist-sans",
+  subsets: ["latin"],
+});
+
+const geistMono = Geist_Mono({
+  variable: "--font-geist-mono",
+  subsets: ["latin"],
+});
+
+const roboto = Roboto({
+  variable: "--font-roboto-sans",
+  subsets: ['latin'],
+});
+
+export default function MyApp({ Component, pageProps }: AppProps) {
+  return (
+    <>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <style jsx global>{`
+        html {
+          font-family: ${geistSans.style.fontFamily}, ${geistMono.style.fontFamily}, ${roboto.style.fontFamily};
+        }
+      `}</style>
+      <I18nProvider>
+        <Component {...pageProps} />
+      </I18nProvider>
+    </>
+  );
+}
