+++ tsconfig.json
@@ -19,9 +19,8 @@
       }
     ],
     "paths": {
-      "@/*": ["./src/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/pages", "src/components", "src/types"],
   "exclude": ["node_modules"]
 }
